# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AssessmentSessionInputBand = typeof AssessmentSessionInputBand[keyof typeof AssessmentSessionInputBand];


export const AssessmentSessionInputBand = {
  low: 'low',
  moderate: 'moderate',
  high: 'high',
} as const;

export type AssessmentSessionInputCategoryScores = {
  /**
     * @minimum 0
     * @maximum 100
     */
  emotional: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  physical: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  behavioral: number;
};

export type AssessmentSessionInputAnswersItem = {
  questionId: string;
  /**
     * @minimum 0
     * @maximum 4
     */
  answerValue: number;
};

export interface AssessmentSessionInput {
  /**
     * @minimum 0
     * @maximum 100
     */
  mcqScore: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  finalScore: number;
  band: AssessmentSessionInputBand;
  categoryScores: AssessmentSessionInputCategoryScores;
  facialDataUsed: boolean;
  /**
     * @minimum 0
     * @maximum 100
     * @nullable
     */
  facialTensionIndex?: number | null;
  crisisFlag: boolean;
  answers?: AssessmentSessionInputAnswersItem[];
}

export type AssessmentSession = AssessmentSessionInput & {
  id: number;
  createdAt: string;
};