# lib/api-zod/src/generated/types/assessmentSession.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentSessionInput } from './assessmentSessionInput';

export type AssessmentSession = AssessmentSessionInput & {
  id: number;
  createdAt: Date;
};