# lib/api-zod/src/generated/types/assessmentSessionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentSessionInputAnswersItem } from './assessmentSessionInputAnswersItem';
import type { AssessmentSessionInputBand } from './assessmentSessionInputBand';
import type { AssessmentSessionInputCategoryScores } from './assessmentSessionInputCategoryScores';

export interface AssessmentSessionInput {
  /**
     * @minimum 0
     * @maximum 100
     */
  mcqScore: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  finalScore: number;
  band: AssessmentSessionInputBand;
  categoryScores: AssessmentSessionInputCategoryScores;
  facialDataUsed: boolean;
  /**
     * @minimum 0
     * @maximum 100
     * @nullable
     */
  facialTensionIndex?: number | null;
  crisisFlag: boolean; journalEntry?: string | null;
  answers?: AssessmentSessionInputAnswersItem[];
}